+++ addOns/scripts/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Provide the `script` API on newer ZAP versions.
 
+### Fixed
+- Handle missing "references" field in the script metadata correctly.
+
 ## [45.4.0] - 2024-05-16
 ### Added
 - Support for Automation Framework loaddir action, which loads all of the scripts under the specified directory.